backend/requirements.txt
# Core FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# CORS middleware - use built-in FastAPI middleware

# AI/ML Libraries
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
pinecone-client==2.2.4
transformers==4.35.0
torch==2.1.0

# Vector embeddings - Google EmbeddingGemma
sentence-transformers>=2.2.2
numpy>=1.24.3
pandas>=2.0.3
# For Google EmbeddingGemma support
hf_hub>=0.20.0

# Async database support
aiosqlite==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0